Added second integration test
